Fixed -Wsign-compare warnings
unit/test-sailfish_sim_info.c:86:33: signed and unsigned type in conditional expression
drivers/ril/ril_netmon.c:46:12: comparison between signed and unsigned integer expressions

+++ ofono/unit/test-sailfish_sim_info.c
@@ -83,7 +83,7 @@ struct ofono_netreg {
 
 int ofono_netreg_get_status(struct ofono_netreg *netreg)
 {
-	return netreg ? netreg->status : -1;
+	return netreg ? (int) netreg->status : -1;
 }
 
 const char *ofono_netreg_get_mcc(struct ofono_netreg *netreg)
